Move coverage.js into test/unit
